+++ src/main/java/com/bank/service/account/AccountAmountService.java
@@ -4,8 +4,7 @@ import com.bank.bean.account.AccountBean;
 import com.bank.bean.acountsavings.AccountSavingBean;
 import com.bank.exception.InvalidParamValueException;
 import com.bank.projection.account.AccountAmountProjection;
-import com.bank.repository.account.AccountRepository;
-import com.bank.service.accountsaving.AccountSavingService;
+import com.bank.service.account.accountsaving.AccountSavingService;
 import com.bank.util.Logging.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,10 +21,10 @@ public class AccountAmountService {
         Logger.info("Getting balance of accountId:%s",accountId);
         AccountBean accountBean = accountService.getAccountBeanByAccountId(accountId);
         AccountAmountProjection projection = new AccountAmountProjection();
-        projection.setBalance(accountBean.getAmount());
+        projection.setBalance(accountBean.getAmount().doubleValue());
         try {
             AccountSavingBean accountSavingBean = accountSavingService.getAccountSavingsBeanByAccountBean(accountBean);
-            projection.setSavingAccountBalance(accountSavingBean.getAmount());
+            projection.setSavingAccountBalance(accountSavingBean.getAmount().doubleValue());
         }catch (InvalidParamValueException e){
             projection.setSavingAccountBalance(0);
         }
